pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image-share"
version = "0.1.0"
authors = [
{ name="Alex Pope", email="alex@aroundwithalex.dev" }
]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.6.2.post1",
    "fastapi[standard]>=0.115.2",
    "httpx>=0.27.2",
    "pydantic>=2.9.2",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.13",
    "scikit-learn>=1.5.2",
    "sqlalchemy>=2.0.36",
    "trio>=0.27.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating Sytem :: OS Independent"
]

[tool.uv]
dev-dependencies = [
    "nox>=2024.10.9",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
]
